Change smapi command to cp command in zvm-agent
If zvm-agent is not on the same zvm with xcat,when zvm-agent use smapi command to query xcat info from its zhcp,
it will get an error result, becuse the result it get is the xcat info on same zvm.
so change these command from SMCLI to CP or linux commands.

Change-Id: I38cea9394f76b90073cef459a6d00a29422178e3