Make enhancement for zvm-agent log for tracking issue more easily
When the commands in xcat restapi executed result is not successful,
the result xcat restapi will contain the result of the command detail
and we catch the error and append the error detail information to log file.

Remove the '/' from tox.ini copyfile directory, because of copy failed on mac os.

Change-Id: Ieac42491bb41067e0efe980696035043288fa522