Performance enhance to remove couple operation
Reduce one smcli call for user direct to improve performance.

The user direct for NIC will be updated until all node info updated
in nova and neutron-agent side, we don't need to create a
NIC , then assign a vswitch to it, we can do this in a whole bunch.

This patch should work with nova's patch https://review.openstack.org/306947

Change-Id: I714783c2eda42f37e90fe9a093fed546a31cf205