Use https with CA conf
CA file is a new added config introduced to be used by neutron-zvm-agent
service. When compute service needs to communicate with xcat, it will add
the CA file into https communication so it can verify the xcat is the one
that neutron-zvm-agent wants to communicate with.

see following link for more info.
https://bugs.launchpad.net/ossn/+bug/1188189

Change-Id: I66964ca36b4b80a6b68ee52db5d06a652ec18f10